+++ app/desktop/src/ipc.ts
@@ -1,4 +1,4 @@
-import type { DocumentUpdate, TerminalMessage } from "@core/api";
+import type { DocumentUpdate, Evaluation, Log } from "@core/api";
 
 import type { SavedStatus } from "./main/filesystem";
 
@@ -23,8 +23,9 @@ export interface ToRendererChannels {
   status: { withID: string; content: SavedStatus };
   setCurrent: { id: string };
   close: { id: string };
-  console: TerminalMessage;
+  console: Evaluation | Log;
   tidalVersion: string;
   tidalNow: number;
+  toggleConsole: undefined;
   showAbout: string;
 }