+++ app/desktop/src/main/index.ts
@@ -33,6 +33,7 @@ const createWindow = (
     height: 600,
     webPreferences: {
       preload: resolve(app.getAppPath(), "build/preload/index.js"),
+      sandbox: process.env.NODE_ENV === "production",
     },
   });
 
@@ -103,36 +104,7 @@ const createWindow = (
 
     listeners.push(
       listen("requestClose", async ({ id }) => {
-        let document = filesystem.getDoc(id);
-
-        if (!document) throw Error("Tried to close a non-existent document");
-
-        if (!document.saved) {
-          let { response } = await dialog.showMessageBox(window, {
-            type: "warning",
-            message: "Do you want to save your changes?",
-            buttons: ["Save", "Don't Save", "Cancel"],
-          });
-
-          // Cancelled
-          if (response === 2) return;
-
-          // Save
-          if (response === 0) {
-            if (document.path) {
-              document.save();
-            } else {
-              let { canceled, filePath } = await dialog.showSaveDialog(window);
-
-              if (!canceled && filePath) {
-                document.save(filePath);
-              }
-            }
-          }
-        }
-
-        // We're done here, so close the file
-        send("close", { id });
+        await close({ window, id });
       })
     );
 
@@ -150,6 +122,12 @@ const createWindow = (
     listeners.push(
       menu.on("rebootTidal", () => {
         tidal.restart();
+      })
+    );
+
+    listeners.push(
+      menu.on("toggleConsole", () => {
+        send("toggleConsole", undefined);
       })
     );
 
@@ -326,6 +304,9 @@ async function close({ window, id }: CloseOptions) {
       }
     }
   }
+
+  // Close document
+  await document.close();
 
   // We're done here, so close the file
   send("close", { id });