+++ app/desktop/src/preload/index.ts
@@ -82,6 +82,8 @@ const ElectronAPI = {
 
   onConsoleMessage: listen("console"),
 
+  onToggleConsole: listen("toggleConsole"),
+
   onTidalVersion: listen("tidalVersion"),
 
   onTidalNow: listen("tidalNow"),