+++ app/desktop/src/renderer/main.ts
@@ -11,12 +11,11 @@ import { settings } from "@core/extensions/settings/editor";
 import { LayoutView } from "@core/extensions/layout";
 import { console as electronConsole } from "@core/extensions/console";
 // import { peer } from "@core/extensions/peer";
-import { toolbar } from "@core/extensions/toolbar";
+import { toolbarConstructor } from "@core/extensions/toolbar";
 
 import { fileSync } from "./file";
 import { EditorTabView } from "@core/extensions/layout/tabs/editor";
 import { AboutTabView } from "@core/extensions/layout/tabs/about";
-import { ConsoleMessage } from "packages/codemirror/console/src";
 
 window.addEventListener("load", () => {
   const parent = document.body.appendChild(document.createElement("section"));
@@ -43,14 +42,24 @@ export class Editor {
 
     // Keep track of Tidal state
     let tidalVersion: string | undefined;
-    let tidalConsole: ConsoleMessage[] = [];
+
+    // Append Tidal UI Panels
+    let tidalConsole = electronConsole();
+    layout.panelArea.appendChild(tidalConsole.dom);
+
+    let toolbar = toolbarConstructor(api, tidalVersion);
+    layout.panelArea.appendChild(toolbar.dom);
 
     api.onTidalVersion((version) => {
       tidalVersion = version;
     });
 
+    api.onToggleConsole(() => {
+      tidalConsole.toggleVisibility();
+    });
+
     api.onConsoleMessage((message) => {
-      tidalConsole.push(message);
+      tidalConsole.update(message);
     });
 
     api.onOpen(({ id, path }) => {
@@ -75,8 +84,6 @@ export class Editor {
                     { path, saved, version, thisVersion: version },
                     api
                   ),
-                  electronConsole(api, tidalConsole),
-                  toolbar(api, tidalVersion),
                   // peer(version),
                 ],
               }),