+++ core/extensions/console/index.ts
@@ -1,31 +1,62 @@
-import { keymap, ViewPlugin } from "@codemirror/view";
-import {
-  ConsoleMessage,
-  consoleState,
-  sendToConsole,
-  clearConsole,
-  console as rootConsole,
-} from "@management/cm-console";
+import { Evaluation, Log } from "@core/api";
 
-import { ElectronAPI } from "@core/api";
+import { messageConstructor } from "./message";
+import "./style.css";
 
-export function console(api: typeof ElectronAPI, initial: ConsoleMessage[]) {
-  const consoleListener = ViewPlugin.define((view) => {
-    const unlisten = api.onConsoleMessage((message) => {
-      view.dispatch(sendToConsole(view.state, message));
-    });
+export function console(history: (Evaluation | Log)[] = []) {
+  let consoleNode = document.createElement("div");
+  consoleNode.classList.add("cm-console");
 
-    return {
-      destroy: () => {
-        unlisten();
-      },
-    };
+  consoleNode.setAttribute("role", "log");
+  consoleNode.tabIndex = 0;
+
+  let consoleWrapperNode = document.createElement("div");
+  consoleNode.appendChild(consoleWrapperNode);
+
+  let consoleHeight = 0;
+  let wrapperHeight = 0;
+
+  const observer = new ResizeObserver((elements) => {
+    for (let element of elements) {
+      if (element.target === consoleNode) {
+        consoleHeight = element.contentBoxSize[0].blockSize;
+      } else if (element.target === consoleWrapperNode) {
+        wrapperHeight = element.contentBoxSize[0].blockSize;
+      }
+    }
+
+    consoleWrapperNode.classList.toggle(
+      "scrolling",
+      wrapperHeight > consoleHeight
+    );
   });
 
-  return [
-    consoleState.init(() => [...initial]),
-    consoleListener,
-    rootConsole(),
-    keymap.of([{ key: "Mod-`", run: clearConsole }]),
-  ];
+  observer.observe(consoleNode);
+  observer.observe(consoleWrapperNode);
+
+  for (let message of history) {
+    consoleWrapperNode.appendChild(messageConstructor(message));
+  }
+
+  let visible = true;
+
+  const toggleVisibility = (value?: boolean) => {
+    visible = value ?? !visible;
+
+    consoleNode.style.display = visible ? "inherit" : "none";
+  };
+
+  return {
+    dom: consoleNode,
+    update(message: Evaluation | Log) {
+      let lastElement = consoleWrapperNode.appendChild(
+        messageConstructor(message)
+      );
+
+      toggleVisibility(true);
+      lastElement.scrollIntoView({ behavior: "smooth" });
+    },
+    toggleVisibility,
+    destroy() {},
+  };
 }