+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Console and toolbar UI are now persistent across editor tabs (#65)
+
 ## [0.9.0] - 2023-10-15
 
 ### Added
@@ -74,6 +80,7 @@ of text.management. They were never tagged in the repository, and NPM updates
 have been discontinued for the time being as I focus on an Electron/web app
 distributed through GitHub releases.
 
+[Unreleased]: https://github.com/mindofmatthew/text.management/compare/v0.9.0...HEAD
 [0.9.0]: https://github.com/mindofmatthew/text.management/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/mindofmatthew/text.management/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/mindofmatthew/text.management/compare/v0.6.0...v0.7.0